+++ SwiftyStoreKit/SKProduct+LocalizedPrice.swift
@@ -28,9 +28,13 @@ import StoreKit
 public extension SKProduct {
 
     public var localizedPrice: String? {
-        let numberFormatter = NumberFormatter()
-        numberFormatter.locale = self.priceLocale
-        numberFormatter.numberStyle = .currency
-        return numberFormatter.string(from: self.price)
+        return priceFormatter(locale: priceLocale).string(from: price)
+    }
+    
+    private func priceFormatter(locale: Locale) -> NumberFormatter {
+        let formatter = NumberFormatter()
+        formatter.locale = locale
+        formatter.numberStyle = .currency
+        return formatter
     }
 }