+++ SwiftyStoreKit/SwiftyStoreKit+Types.swift
@@ -26,7 +26,7 @@ import StoreKit
 
 // MARK: Purchases
 
-// Purchased or restored product
+// Restored product
 public struct Purchase {
     public let productId: String
     public let quantity: Int
@@ -35,11 +35,13 @@ public struct Purchase {
     public let needsFinishTransaction: Bool
 }
 
+// Purchased product
 public struct PurchaseDetails {
     public let productId: String
     public let quantity: Int
     public let product: SKProduct
     public let transaction: PaymentTransaction
+    public let originalTransaction: PaymentTransaction?
     public let needsFinishTransaction: Bool
 }
 
@@ -50,6 +52,7 @@ public protocol ReceiptValidator {
 
 // Payment transaction
 public protocol PaymentTransaction {
+    var transactionDate: Date? { get }
     var transactionState: SKPaymentTransactionState { get }
     var transactionIdentifier: String? { get }
 }